+++ CHANGELOG.md
@@ -16,6 +16,7 @@ branch is the latest unreleased version.
 - Implement face fields on Axes, matching Roblox API.
 - Add `--include-root` flag to run command to include paths as root directories.
 - Add `--allow-insecure-paths` flag to run command to disable path restrictions.
+- Add CFrame.lookAt constructor.
 
 **Fixes:**
 - Fix version displayed by rbxmk.